requirements.txt
ipykernel
ipython
pandas
scipy
requests
gymnasium
ray[default]==2.1.0
ray[rllib]
ray[serve]
ray[tune]
optuna
locust
mlflow
torch
torchmetrics
pytorch-lightning==1.8.4
scikit-learn
joblib
tqdm
seaborn
tensorboard
pyarrow
pyyaml
azure-core
azure-keyvault-secrets
azure-identity
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
uamqp
azure-ai-ml
azureml-mlflow
azureml-inference-server-http
databricks-sql-connector
